chore: remove roller code for electron <=3
